change of profile (as for commit 16731... I have been to fast (one more time)

+++ scilab/modules/core/src/c/basout.h
@@ -6,6 +6,6 @@
 #define __BANNER_H__
 
 #include "machine.h"
-int C2F(basout)(integer *io,integer *lunit,char *string,long int notused);
+int C2F(basout)(integer *io,integer *lunit,char *string,int notused);
 
 #endif /* __BANNER_H__ */
